Tarefa0_2025

Description

Módulo que define funções auxiliares que serão úteis na resolução do trabalho prático de LI1/LP1 em 2025/26.

Documentation

encontraQuantidadeArmaMinhoca :: TipoArma -> Minhoca -> Int

Retorna a quantidade de munições disponíveis de uma minhoca para uma dada arma.

atualizaQuantidadeArmaMinhoca :: TipoArma -> Minhoca -> Int -> Minhoca

Atualiza a quantidade de munições disponíveis de uma minhoca para uma dada arma.

eTerrenoDestrutivel :: Terreno -> Bool

Verifica se um tipo de terreno é destrutível, i.e., pode ser destruído por explosões.

NB: Apenas Terra é destrutível.

eTerrenoOpaco :: Terreno -> Bool

Verifica se um tipo de terreno é opaco, i.e., não permite que objetos ou minhocas se encontrem por cima dele.

NB: Apenas Terra ou Pedra são opacos.

ePosicaoMapaLivre :: Posicao -> Mapa -> Bool

Verifica se uma posição do mapa está livre, i.e., pode ser ocupada por um objeto ou minhoca.

NB: Uma posição está livre se não contiver um terreno opaco.

ePosicaoEstadoLivre :: Posicao -> Estado -> Bool

Verifica se uma posição do estado está livre, i.e., pode ser ocupada por um objeto ou minhoca.

NB: Uma posição está livre se o mapa estiver livre e se não estiver já uma minhoca ou um barril nessa posição.

destroiPosicao :: Posicao -> Mapa -> Mapa

Destrói uma dada posição no mapa (tipicamente como consequência de uma explosão).

NB: Só terrenos Terra pode ser destruídos. Assumo que seja destruir e substituir por Ar

adicionaObjeto :: Objeto -> Estado -> Estado

Adiciona um novo objeto a um estado.

NB: A posição onde é inserido não é relevante.

Auxiliares

encontraLinhaMapa :: Int -> [Terreno] -> Bool

Verifica se numa linha do Mapa a Posicao dada é livre (não Opaca).(Utilizado na função ePosicaoMapaLivre)

existeMinhoca :: Posicao -> [Minhoca] -> Bool

Verifica se numa lista de minhocas já existe uma Minhoca na dada Posicao.(Utilizado na função ePosicaoEstadoLivre)

existeBarril :: Posicao -> [Objeto] -> Bool

Verifica se numa lista de objetos já existe um Barril na dada Posicao.(Utilizado na função ePosicaoEstadoLivre)

existeObjeto :: Posicao -> [Objeto] -> Maybe Objeto

ehDisparo :: Objeto -> Bool

Verifica se um Objeto é um Disparo.(Utilizado na função existeBarril)

posicaoObjeto :: Objeto -> Posicao

Devolve a Posição de um Objeto.(Utilizado na função existeBarril)

minhocaTemDisparo :: TipoArma -> NumMinhoca -> [Objeto] -> Bool

Verifica se numa lista de objetos já existe um Disparo feito para uma dada arma(TipoArma) por uma dada minhoca(NumMinhoca).

verificaVida :: Minhoca -> Bool

Verifica se a vida da minhoca é valida

Exemplos:

Expand

>>> verificaVida Minhoca{posicaoMinhoca=Just (3,0), vidaMinhoca=Morta, jetpackMinhoca=100, escavadoraMinhoca=200, bazucaMinhoca=150, minaMinhoca=3, dinamiteMinhoca=1}
True
>>> verificaVida Minhoca{posicaoMinhoca=Just (3,0), vidaMinhoca=Viva 150, jetpackMinhoca=100, escavadoraMinhoca=200, bazucaMinhoca=150, minaMinhoca=3, dinamiteMinhoca=1}
False

eMinhocaViva :: Minhoca -> Bool

Verifica se uma minhoca esta viva