Tarefa1

Description

Módulo para a realização da Tarefa 1 de LI1/LP1 em 2025/26.

Documentation

validaEstado :: Estado -> Bool

Função principal da Tarefa 1. Recebe um Estado e retorna se este é válido ou não.

Exemplo de Utilização:

Expand

Um exemplo de estado válido seria:

estadoValido = Estado
    { mapaEstado =
        [[Ar,Ar,Ar,Ar,Ar,Ar]
        ,[Ar,Ar,Ar,Ar,Ar,Ar]
        ,[Terra,Terra,Terra,Pedra,Agua,Agua]
        ,[Terra,Terra,Terra,Terra,Pedra,Agua]
        ]
    , objetosEstado =[]
    , minhocasEstado =
        [Minhoca {posicaoMinhoca = Just (1,1), vidaMinhoca = Viva 100, jetpackMinhoca = 1, escavadoraMinhoca = 1, bazucaMinhoca = 1, minaMinhoca = 1, dinamiteMinhoca = 1}]
    }

Ao validar o estado anterior, a função retorna True

>>> validaEstado estadoValido
True

Funções Auxiliares

eMapaValido :: Mapa -> Bool

A função eMapaValido valida se o Mapa é estruturalmente coerente.

Funcionalidade:

Verifica se o Mapa é uma matriz válida, ou seja, todas as linhas têm o mesmo número de colunas.

Retorna False se o Mapa estiver vazio.

eMapaValido [] = False    
eMapaValido m = (eMatrizValida m) 

Exemplos de Utilização:

Expand

>>> eMapaValido [[Ar,Ar,Ar],[Ar,Ar,Ar],[Ar,Ar,Ar]]
True
>>> eMapaInvalido [[Ar,Ar,Terra],[Ar,Ar],[Ar,Ar,Terra]]
False

eObjetosValido :: Estado -> [Objeto] -> Bool

A função eMapaValido valida se os objetos de um estado são consistentes. Verifica disparos, posições válidas e coerência de donos.

Exemplos:

Expand

mapaValido =
        [[Ar,Ar,Ar,Ar,Ar,Ar]
        ,[Ar,Ar,Ar,Ar,Ar,Ar]
        ,[Terra,Terra,Terra,Pedra,Agua,Agua]
        ,[Terra,Terra,Terra,Terra,Pedra,Agua]
        ]
objetoValido = Disparo{posicaoDisparo=(1,4), direcaoDisparo=Norte, tipoDisparo=Bazuca, tempoDisparo=Nothing, donoDisparo=0}
minhocaValida = Minhoca{posicaoMinhoca=Just (1,1), vidaMinhoca=Viva 100, jetpackMinhoca=1, escavadoraMinhoca=1, bazucaMinhoca=1, minaMinhoca=1, dinamiteMinhoca=1}

>>> eObjetosValido Estado{mapaEstado=mapaValido, objetosEstado=[objetoValido], minhocasEstado=[minhocaValida]} [objetoValido]
True

disparoValido :: Mapa -> Objeto -> Bool

Verifica se o Disparo é valido

Funcionalidade:

Verifica se o tempo do disparo é valido(tempoDisparo)

Verifica se a Posicao do disparo é valida

Verifica se o dono do disparo é valido

Verifica se o disparo nao colide com outro Objeto ou Minhoca(so para Mina e Dinamite)

Exemplos:

Expand

mapaValido =
        [[Ar,Ar,Ar,Ar,Ar,Ar]
        ,[Ar,Ar,Ar,Ar,Ar,Ar]
        ,[Terra,Terra,Terra,Pedra,Agua,Agua]
        ,[Terra,Terra,Terra,Terra,Pedra,Agua]
        ]

>>> disparoValido mapaValido Disparo{posicaoDisparo=(1,4), direcaoDisparo=Norte, tipoDisparo=Bazuca, tempoDisparo=Nothing, donoDisparo=0}
True

disparoBazucaValido :: Objeto -> Mapa -> Bool

Verifica se o Disparo de Bazuca é valido

Funcionalidade:

Verifica se o bloco anterior ao Disparo é livre (não opaco)

Verifica se o disparo não colide com outro Objeto ou Minhoca

Exemplos:

Expand

>>> disparoBazucaValido Disparo{posicaoDisparo=(3,4), direcaoDisparo=Norte, tipoDisparo=Bazuca, tempoDisparo=Nothing, donoDisparo=0} mapaValido
False
>>> disparoBazucaValido Disparo{posicaoDisparo=(1,4), direcaoDisparo=Norte, tipoDisparo=Bazuca, tempoDisparo=Nothing, donoDisparo=0} mapaValido
True

Verificação dos Donos

donoValido :: Estado -> Objeto -> Bool

Verifica se o dono do Disparo é valido

Funcionalidade:

Verifica se o indice do dono é valido na lista de minhocas

Verifica se o dono não tem mais do que um disparo do mesmo tipo

Verifica se o dono existe na lista de minhocas do estado

Exemplos:

Expand

estado1 = Estado { mapaEstado = [[Ar]], 
                   objetosEstado = [Disparo {posicaoDisparo = (1,1), 
                   direcaoDisparo = Norte, tipoDisparo = Mina, 
                   tempoDisparo = Just 2, donoDisparo = 0}], 
                   minhocasEstado = [Minhoca {posicaoMinhoca = Just (1,1), 
                   vidaMinhoca = Viva 100, 
                   jetpackMinhoca = 1, escavadoraMinhoca = 1, bazucaMinhoca = 1, minaMinhoca = 1, dinamiteMinhoca = 1}] }   

>>> donoValido estado1 (head (objetosEstado estado1))
True

listaDonos :: [Objeto] -> [(TipoArma, Int)]

Função auxiliar para donoValido. Devolve uma lista com o tipo de arma e o dono(NumMinhoca) de cada Disparo

Funcionalidade:

Cria uma lista de tuplos (TipoArma, NumMinhoca) a partir de uma lista de objetos

Cada tuplo representa o tipo de arma e o dono do disparo correspondente

Exemplos:

Expand

>>> listaDonos [Disparo{posicaoDisparo=(1,4), direcaoDisparo=Norte, tipoDisparo=Mina, tempoDisparo=Just 2, donoDisparo=0}, Disparo{posicaoDisparo=(2,3), direcaoDisparo=Sul, tipoDisparo=Dinamite, tempoDisparo=Just 3, donoDisparo=1}]
[(Mina,0),(Dinamite,1)]

eMinhocasValidas :: Estado -> [Minhoca] -> Bool

Valida se as minhocas são validas

Funcionalidade:

Verifica se a Posicao da minhoca é valida

Verifica se a morte da minhoca é valida(validaMorte)

Verifica se a vida da minhoca é valida(verificaVida)

Verifica se as armas da minhoca são validas

Exemplos:

Expand

estadoValido = Estado
    { mapaEstado =
        [[Ar,Ar,Ar,Ar,Ar,Ar]
        ,[Ar,Ar,Ar,Ar,Ar,Ar]
        ,[Terra,Terra,Terra,Pedra,Agua,Agua]
        ,[Terra,Terra,Terra,Terra,Pedra,Agua]
        ]
    , objetosEstado =[]
    , minhocasEstado =
        [Minhoca {posicaoMinhoca = Just (1,1), vidaMinhoca = Viva 100, jetpackMinhoca = 1, escavadoraMinhoca = 1, bazucaMinhoca = 1, minaMinhoca = 1, dinamiteMinhoca = 1}]
    }

>>> eMinhocasValidas estadoValido (minhocasEstado estadoValido)
True

>>> eMinhocasValidas estadoInvalido (minhocasEstado estadoInvalido)
False

Auxiliares Minhoca

validaMorte :: Minhoca -> Mapa -> Bool

Verifica se a morte da minhoca é valida

Funcionalidade:

Verifica se a Posicao da minhoca é Nothing e se a vida é Morta

Verifica a Posicao da minhoca e se o terreno nessa posição é Agua e se a vida é Morta

Exemplos:

Expand

>>> validaMorte Minhoca{posicaoMinhoca=Just (5,8), vidaMinhoca=Morta, jetpackMinhoca=100, escavadoraMinhoca=200, bazucaMinhoca=150, minaMinhoca=3, dinamiteMinhoca=1} mapaValido
True